NFH-main/FInalNFH/Honeypot/requirements.txt
scapy
python-magic
selenium==4.1.0
pyftpdlib
dhcppython
dnspython
